requirements.txt
streamlit
requests
beautifulsoup4
pandas
googlemaps
google-generativeai
sendgrid
pyperclip